tests/wpt/tests/css/css-display/reading-order/reading-order-items-valid.html

<!DOCTYPE html>
<title>CSS Display: parsing reading-order-items with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-display-4/#reading-order-items">
<link rel="author" title="Di Zhang" href="mailto:dizhangg@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<body>
<script>
test_valid_value('reading-order-items', 'normal');
test_valid_value('reading-order-items', 'flex-visual');
test_valid_value('reading-order-items', 'flex-flow');
test_valid_value('reading-order-items', 'grid-rows');
test_valid_value('reading-order-items', 'grid-columns');
test_valid_value('reading-order-items', 'grid-order');
</script>
</body>
